Update Prisma Schema
Rename model RSVP to Rsvp to follow Prisma automatic naming conventions.
Set Host and Rsvp to cascade on delete, i.e. if an Event is deleted, all
related Hosts and Rsvps are also deleted. If a User is deleted, all related
Rsvps are also deleted. Additional logic is needed for Hosts associated with
a User.