Create new layouts folder and combine SignInFlow
SignInForm component now lives inside SignInFlow file. There was no need to
have them be separate. SignInForm can still be imported without SignInFlow.

Any component that looks like a page are to be located in the layouts folder.
SignInFlow is an exception to this. I don't actually have a good reason why
right now, but it's the only thing rendered on the signin route besides the
Card containing it.

Most of the components called by route files now live in the new layouts folder
(the exception to that are components that are used to navigate between
layouts).